fix: sonar Duplicated Lines 이슈 해결
mock 데이터로 이슈 수정
