+++ frontend/src/mock/submissions.ts
@@ -1,0 +1,60 @@
+const submissions = [
+  {
+    submissionId: 1,
+    jobId: 1,
+    jobName: '청소',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 2,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 3,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 4,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 5,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 6,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 7,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+  {
+    submissionId: 8,
+    jobId: 2,
+    jobName: '마감',
+    author: '어썸오',
+    createdAt: 'YYYY-MM-DDTHH:MM:SS',
+  },
+];
+
+export default submissions;
