+++ frontend/src/pages/host/DashBoard/index.tsx
@@ -1,67 +1,9 @@
+import submissions from '@/mock/submissions';
 import { css } from '@emotion/react';
 import { useNavigate } from 'react-router-dom';
 
 import JobList from '@/components/host/JobList';
 import Submissions from '@/components/host/Submissions';
-
-const submissions = [
-  {
-    submissionId: 1,
-    jobId: 1,
-    jobName: '청소',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 2,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 3,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 4,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 5,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 6,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 7,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-  {
-    submissionId: 8,
-    jobId: 2,
-    jobName: '마감',
-    author: '어썸오',
-    createdAt: 'YYYY-MM-DDTHH:MM:SS',
-  },
-];
 
 const DashBoard: React.FC = () => {
   const navigate = useNavigate();
